// src/app/api/model/messageDto.ts
/**
 * TaskFlow Chat API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MessageContentType } from './messageContentType';
import { MessageMetadata } from './messageMetadata';
import { MessageSourceType } from './messageSourceType';
import { MessageType } from './messageType';


export interface MessageDto { 
    messageId?: string;
    groupId?: string;
    senderId?: string | null;
    senderName?: string | null;
    sourceType?: MessageSourceType;
    messageType?: MessageType;
    contentType?: MessageContentType;
    content?: object;
    metadata?: MessageMetadata;
    createdAt?: string;
}
export namespace MessageDto {
}