// src/app/api/model/sendMessageDto.ts
/**
 * TaskFlow Chat API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MessageContentType } from './messageContentType';
import { MessageMetadata } from './messageMetadata';
import { MessageType } from './messageType';


export interface SendMessageDto { 
    messageType?: MessageType;
    contentType?: MessageContentType;
    content?: object;
    metadata?: MessageMetadata;
}
export namespace SendMessageDto {
}